+++ includes/class-bootstrap.php
@@ -23,7 +23,6 @@ namespace Newspack_Nodes;
 use Newspack_Nodes\Rest\Command_Controller;
 use Newspack_Nodes\Rest\Messages_Stream_Controller;
 use Newspack_Nodes\Rest\SpawnController;
-use Newspack_Nodes\Rest\TopologyStreamController;
 
 \defined( 'ABSPATH' ) || exit;
 
@@ -151,7 +150,6 @@ class Bootstrap {
 	 */
 	public static function register_rest_routes(): void {
 		( new SpawnController( self::supervisor() ) )->register_routes();
-		( new TopologyStreamController() )->register_routes();
 		( new Messages_Stream_Controller() )->register_routes();
 		( new Command_Controller() )->register_routes();
 	}